🦄 refactor(RepoList.tsx): Split RepoList
Make RepoList cleaner and splitted into smaller components, external hooks and interfaces
