+++ src/components/repo/types.ts
@@ -1,0 +1,17 @@
+export interface ApiRepository {
+    id: number;
+    name: string;
+    language: string;
+    stargazers_count: number;
+}
+
+export interface RepositoryListProps {
+    username: string;
+}
+
+export interface Repository {
+    id: number;
+    name: string;
+    language: string;
+    stars: number;
+}
